@@ -1,7 +1,7 @@
 [package]
 name = "zed_elixir"
 version = "0.2.4"
-edition = "2021"
+edition = "2024"
 publish = false
 license = "Apache-2.0"
 
