@@ -1,9 +1,10 @@
 use std::fs;
 
-use zed::lsp::{Completion, CompletionKind, Symbol};
-use zed::{CodeLabel, CodeLabelSpan, LanguageServerId};
-use zed_extension_api::settings::LspSettings;
-use zed_extension_api::{self as zed, Result};
+use zed_extension_api::{
+    self as zed, CodeLabel, CodeLabelSpan, LanguageServerId, Result, Worktree,
+    lsp::{Completion, CompletionKind, Symbol},
+    settings::LspSettings,
+};
 
 use crate::language_servers::util;
 
@@ -28,49 +29,67 @@ impl Lexical {
     pub fn language_server_binary(
         &mut self,
         language_server_id: &LanguageServerId,
-        worktree: &zed::Worktree,
+        worktree: &Worktree,
     ) -> Result<LexicalBinary> {
-        let binary_settings = LspSettings::for_worktree("lexical", worktree)
+        let binary_name = format!("{}/bin/start_lexical.sh", Self::LANGUAGE_SERVER_ID);
+        let binary_settings = LspSettings::for_worktree(Self::LANGUAGE_SERVER_ID, worktree)
             .ok()
             .and_then(|lsp_settings| lsp_settings.binary);
         let binary_args = binary_settings
             .as_ref()
             .and_then(|binary_settings| binary_settings.arguments.clone());
+        let port_wrapper = format!("{}/priv/port_wrapper.sh", Self::LANGUAGE_SERVER_ID);
+        let debug_shell = format!("{}/bin/debug_shell.sh", Self::LANGUAGE_SERVER_ID);
 
-        if let Some(path) = binary_settings.and_then(|binary_settings| binary_settings.path) {
+        if let Some(binary_path) = binary_settings.and_then(|binary_settings| binary_settings.path)
+        {
             return Ok(LexicalBinary {
-                path,
+                path: binary_path,
                 args: binary_args,
             });
         }
 
-        if let Some(path) = worktree.which("lexical") {
+        if let Some(binary_path) = worktree.which(Self::LANGUAGE_SERVER_ID) {
             return Ok(LexicalBinary {
-                path,
+                path: binary_path,
                 args: binary_args,
             });
         }
 
-        if let Some(path) = &self.cached_binary_path {
-            if fs::metadata(path).map_or(false, |stat| stat.is_file()) {
-                return Ok(LexicalBinary {
-                    path: path.clone(),
-                    args: binary_args,
-                });
-            }
+        if let Some(binary_path) = &self.cached_binary_path
+            && fs::metadata(binary_path).is_ok_and(|stat| stat.is_file())
+        {
+            return Ok(LexicalBinary {
+                path: binary_path.clone(),
+                args: binary_args,
+            });
         }
 
         zed::set_language_server_installation_status(
             language_server_id,
             &zed::LanguageServerInstallationStatus::CheckingForUpdate,
         );
-        let release = zed::latest_github_release(
+        let release = match zed::latest_github_release(
             "lexical-lsp/lexical",
             zed::GithubReleaseOptions {
                 require_assets: true,
                 pre_release: false,
             },
-        )?;
+        ) {
+            Ok(release) => release,
+            Err(_) => {
+                if let Some(binary_path) =
+                    util::find_existing_binary(Self::LANGUAGE_SERVER_ID, &binary_name)
+                {
+                    self.cached_binary_path = Some(binary_path.clone());
+                    return Ok(LexicalBinary {
+                        path: binary_path,
+                        args: binary_args,
+                    });
+                }
+                return Err("failed to download latest github release".to_string());
+            }
+        };
 
         let asset_name = format!(
             "{}-{version}.zip",
@@ -85,9 +104,11 @@ impl Lexical {
             .ok_or_else(|| format!("no asset found matching {:?}", asset_name))?;
 
         let version_dir = format!("{}-{}", Self::LANGUAGE_SERVER_ID, release.version);
-        let binary_path = format!("{version_dir}/lexical/bin/start_lexical.sh");
+        let binary_path = format!("{}/{}", version_dir, binary_name);
+        let port_path = format!("{}/{}", version_dir, port_wrapper);
+        let debug_path = format!("{}/{}", version_dir, debug_shell);
 
-        if !fs::metadata(&binary_path).map_or(false, |stat| stat.is_file()) {
+        if !fs::metadata(&binary_path).is_ok_and(|stat| stat.is_file()) {
             zed::set_language_server_installation_status(
                 language_server_id,
                 &zed::LanguageServerInstallationStatus::Downloading,
@@ -101,8 +122,8 @@ impl Lexical {
             .map_err(|e| format!("failed to download file: {e}"))?;
 
             zed::make_file_executable(&binary_path)?;
-            zed::make_file_executable(&format!("{version_dir}/lexical/bin/debug_shell.sh"))?;
-            zed::make_file_executable(&format!("{version_dir}/lexical/priv/port_wrapper.sh"))?;
+            zed::make_file_executable(&port_path)?;
+            zed::make_file_executable(&debug_path)?;
 
             util::remove_outdated_versions(Self::LANGUAGE_SERVER_ID, &version_dir)?;
         }
